+++ crates/world/bin/src/main.rs
@@ -30,21 +30,42 @@ fn main() {
     if let Err(err) =
         Cli::<OpChainSpecParser, WorldChainArgs>::parse().run(|builder, args| async move {
             info!(target: "reth::cli", "Launching node");
-            let node = WorldChainNode::new(args.clone());
-            let handle = builder
-                .node(node)
-                .extend_rpc_modules(move |ctx| {
-                    let provider = ctx.provider().clone();
-                    let pool = ctx.pool().clone();
-                    let sequencer_client =
-                        args.rollup_args.sequencer_http.map(SequencerClient::new);
-                    let eth_api_ext = WorldChainEthApiExt::new(pool, provider, sequencer_client);
-                    ctx.modules.replace_configured(eth_api_ext.into_rpc())?;
-                    Ok(())
-                })
-                .launch()
-                .await?;
-            handle.node_exit_future.await
+
+            if args.flashblock_args.is_some() {
+                let node = WorldChainFlashblocksNode::new(args.clone());
+                let handle = builder
+                    .node(node)
+                    .extend_rpc_modules(move |ctx| {
+                        let provider = ctx.provider().clone();
+                        let pool = ctx.pool().clone();
+                        let sequencer_client =
+                            args.rollup_args.sequencer_http.map(SequencerClient::new);
+                        let eth_api_ext =
+                            WorldChainEthApiExt::new(pool, provider, sequencer_client);
+                        ctx.modules.replace_configured(eth_api_ext.into_rpc())?;
+                        Ok(())
+                    })
+                    .launch()
+                    .await?;
+                handle.node_exit_future.await
+            } else {
+                let node = WorldChainNode::new(args.clone());
+                let handle = builder
+                    .node(node)
+                    .extend_rpc_modules(move |ctx| {
+                        let provider = ctx.provider().clone();
+                        let pool = ctx.pool().clone();
+                        let sequencer_client =
+                            args.rollup_args.sequencer_http.map(SequencerClient::new);
+                        let eth_api_ext =
+                            WorldChainEthApiExt::new(pool, provider, sequencer_client);
+                        ctx.modules.replace_configured(eth_api_ext.into_rpc())?;
+                        Ok(())
+                    })
+                    .launch()
+                    .await?;
+                handle.node_exit_future.await
+            }
         })
     {
         eprintln!("Error: {err:?}");
