+++ crates/world/flashblocks/Cargo.toml
@@ -48,3 +48,4 @@ tracing.workspace = true
 serde_json.workspace = true
 serde.workspace = true
 futures-util = "0.3"
+clap.workspace = true
