+++ crates/world/flashblocks/src/lib.rs
@@ -1,4 +1,3 @@
-pub mod block_building;
+pub mod args;
 pub mod builder;
-pub mod ctx;
 pub mod payload;
