+++ crates/world/node/src/lib.rs
@@ -1,6 +1,7 @@
 #![cfg_attr(not(any(test, feature = "test")), warn(unused_crate_dependencies))]
 
 pub mod args;
+pub mod flashblocks;
 pub mod node;
 
 #[cfg(any(feature = "test", test))]
